Try not using `-static` for macOS linker
It seems like we can use the same flags on Linux and Windows, but macOS
seems to fail when attempting to use `-static` for the linker. We set it
so it doesn't pass the flag when building on macOS.

This might now work, but we're going to commit it and try it in CI. If
it doesn't work, we'll have to try something else. It might be that none
of the `-static` flags will work.