Use extensions explicitly in each module
Up until now, we've had a bunch of language extensions enabled in the
`package.yaml` file. Most of these, we never even used. Our program is
actually relatively simple. We did some hacks to make other tooling
work, like stylish-haskell.

We take an alternative approach here and move al of the language
extensions into the modules. We end up with only using a handful of
extensions per module. This should allow more tooling to work out of the
box without needing to understand `package.yaml` files or be
special-cased.