Move from `stylish-haskell` to `ormolu`
`stylish-haskell` did us well for about two years, but `ormolu` is a bit
more in line with the goals of this project. In particular, it supports
a Vonderhaar-style formatting much like we implemented. A handful of
formatting differences exist, but it's the spirit of the formatter that
matters.

Most of the changes here are whitespace, so it shouldn't matter too much
about what changed.