Move `Verbose` to the top
We want to add a few more modes that the CLI can do. With things setup
as a sum of products, we end up with `optparse-applicative` not knowing
how to deal with the same value in multiple sums. We switch things
around to a product of sums, `optparse-applicative` will work a little
better.