Only validate with `ormolu`
We've had this "lint" step for checking Haskell format. We started with
`stylish-haskell`, and it didn't have a way to validate the format of
files aside from formatting files and checking if something changed. We
switched to `ormolu`, and it allows validating if a file is formatted.
We weren't using this feature of `ormolu` yet, since it seemed that it
could work similar to how `stylish-haskell` was working.

Unfortunately, `ormolu` started changing the file mode of `Setup.hs`
when it would format it. It changed the file mode from 0644 to 0755
starting in this GitLabCI job:
https://gitlab.com/joneshf/purty/-/jobs/446540026. For posterity the
(relavant) output of that run is:

```
$ ci/lint/ormolu.sh --verbose
[DEBUG]      Created temp directory: /tmp/ormolu.sh.kAJ16Ok2nP
[DEBUG]      Created log file: /tmp/ormolu.sh.log.kl5o0XpO8G
[DEBUG]      Building 'ormolu'
[INFO]       Running 'ormolu'
[DEBUG]      Checking for changed files
[ERROR]      Some Haskell files are not formatted properly
[ERROR]      diff --git a/Setup.hs b/Setup.hs
old mode 100644
new mode 100755
[ERROR]
[ERROR]      You can fix this by running 'ci/lint/ormolu.sh' locally and committing
[DEBUG]      Removing temp directory
```

The surprising part of this is that `ormolu` just started doing it. The
previous GitLabCI
job (https://gitlab.com/joneshf/purty/-/jobs/446023209) showed no change
in file mode:
```
$ ci/lint/ormolu.sh --verbose
[DEBUG]      Created temp directory: /tmp/ormolu.sh.NtTGT5y0RG
[DEBUG]      Created log file: /tmp/ormolu.sh.log.bY38hRAShK
[DEBUG]      Building 'ormolu'
[INFO]       Running 'ormolu'
[DEBUG]      Checking for changed files
[DEBUG]      Removing temp directory
```

The commit where that started was
848d407942d832b209474dacea4f431264b7b58a, and that commit has no changes
to anything `ormolu` related. For posterity:

```
$ git show 848d407942d832b209474dacea4f431264b7b58a
commit 848d407942d832b209474dacea4f431264b7b58a
Author: joneshf <jones3.hardy@gmail.com>
Date:   Fri Feb 21 22:01:51 2020 -0800

    Try setting the default python to 3.7

    Looks like `rules_haskell` requires python 3. This appears to be the
    syntax to make python 3 available on the path. Let's see if this works.

diff --git a/appveyor.yml b/appveyor.yml
index ef6b2ee..7fda8ad 100644
--- a/appveyor.yml
+++ b/appveyor.yml
@@ -29,6 +29,7 @@ deploy:
     version: '%APPVEYOR_REPO_TAG_NAME%'

 environment:
+  PYTHON: C:\Python37
   STACK_ROOT: '%APPVEYOR_BUILD_FOLDER%\.stack-root'
   TEST_TMPDIR: '%APPVEYOR_BUILD_FOLDER%\.bazel-cache'
```

It's unclear why this started happening. There's probably some out of
band information we don't know about that's causing this change. We
should report it upstream if it's not known already.

The outcome of this unintentional change is that every GitLabCI job is
now failing. To get around this, we change this "lint" script to
actually lint, instead of formatting and failing if formatting is
different.

We also move the real formatting to the `Makefile`. We do this because
writing another shell file for formatting would be too much boilerplate
for not enough pay off.