Write a failing test for the version
We're going to assume that we'll have a version file for `purty`. We're
doing something a little different than most `cabal`-based projects
would do. We're not depending on `Paths_purty` to figure out the
version. The main reason being: we don't want to couple too hard to
`cabal`.

We're hoping to re-write this in go fairly soon:
https://gitlab.com/joneshf/purty/issues/168. Having another dependency
on a build system for a different language will not be great to try and
fix. Ideally, we'll have an easier time rewriting it if we don't have
extra dependencies.

In any case, we add a failing test so we can see when we're done. We
have two version flags we want to provide: `--numeric-version` which is
useful for machines to process, and `--version` which is useful for
humans to process. We might flip the flags or change something else
about things, but this is what we want to make available.