Support `version` command
We start with the human-readable implementation. This should give us out
some information that's easier to parse and has more context than the
plain version number.

As mentioned, we move the version number out of the `package.yaml`
instead of relying on `cabal` and its ability to generate `Paths_purty`.
There's already one nice thing that came out of this change: we don't
need to pass around the version as an environment variable anymore!

That was a dubious effort to begin with, as we never really knew if we
had the right thing or not. Now, it's tracked in the repo which version
we should be on and what we can do with that.