Add failing test for `format` command
By adding a new command (`version`), we've introduced a breaking change
where we no long will format either a file or directory named "version."
To address this situation, we introduce a new command (`format`) that
allows formatting to occur again even if the file or directory is named
"version."

It's not very likely that someone has a directory named "version" that
they want to format. It's even less likely that someone has a file named
"version" that they want to format. But, it's also not impossible. We
add this alternative way of invoking the binary so we can hopefully
mitigate how much pain people feel.