Download the `hlint` binary
Since `hlint` provides us with a binary, and we don't use the `hlint`
package as a library, we don't need to depend on it through `stack`. We
were only depending on it through `stack` because it was easy to
install. Trying to get up and rolling from scratch takes a non-trivial
amount of time because we're building this tooling from scratch.

This might end up biting us if we don't have all of the dependencies
that `hlint` needs in every environment.