Write a failing test for `validate` mode
We want to first validate that `validate` mode doesn't work. Then we'll
know when we've implemented it the way we want.

We also fix a bug that's been in the way we were testing the exit code
for a while. The string of characters `2&>` don't do what we originally
intended them to do. What we wanted was for
STDERR to be redirected to `/dev/null`. It appears that it puts a `2` as
an argument, forks the process, and redirects STDOUT to `/dev/null`.

We change things around to be more explicit while also fixing this bug.