Build all artifacts with bazel
Now that we've got everything setup correctly, we can build everything
we need with bazel and not worry about using stack for the artifacts.

We still hang on to stack for the `ormolu` and `weeder` rules as
`ormolu` has no pre-built binary we can use instead and `weeder` doesn't
really work unless it's in a stack project. We'll probably be able to
build `ormolu` and run it as a bazel target. And we might end up giving
up on `weeder` for now.

In any case, we're almost taken over everything!