Switch all CI testing over to bazel
The majority of this change is motivated by wanting to keep code
coverage numbers. Currently, there's something in `rules_haskell` that
requires a coverage to be on a package not at the top-level.

We ended up splitting the top-level package into multiple packages.
Since we had to move targets around, we ended up having to fix all the
other things that depended on those targets as well. Not a big deal, but
it means that this commit is larger than necessary.

In any case, we also end up dropping quite a bit of the stack
infrastructure from CI. We might even be able to get rid of most of the
other stack-specific machinery that prevades the repo.