Try referencing directory relative to package
There's something going on with Windows, it doesn't like the directory
referring to the path from the top-level. That's understandable. But,
Linux and macOS don't seem to mind it (those tests run and pass).

This might be flipping the issue, but let's see if it even works on
Windows first.