Simplify `fileUsed` test
Even though there's more code here, the code is actually simpler.
Instead of grabbing all of the files each time and checking for a needle
in a haystack, we grab only the files we want to assert are being used,
and check for their corresponding input file.

We hit the file system a bunch of times, but the complexity has gone
from O(n^2) to O(n). Not a real concern when we have under 1000 files or
so, but it makes other things simpler.

In particular, the reason AppVeyor is failing is because it cannot find
the files it is supposed to have. When we tried to make things work by
looking up the runfiles, it got pretty messy, pretty quick. Hopefully,
by making this test simpler, it will be less messy.