Look up runfiles
Now that we're building with bazel, the location of files we use for
tests is different. It's been working on Linux and macOS because they
work with symlinks, but Windows does not.

We follow the pattern that other languages do and use a
language-specific way of locating the runfiles:
https://docs.google.com/document/d/e/2PACX-1vSDIrFnFvEYhKsCMdGdD40wZRBX3m3aZ5HhVj4CtHPmiXKDCxioTUbYsDydjKtFDAzER5eg7OjJWs3V/pub.

It's unfortunate that we now for sure have some bazel-specific stuff in
our code. There's surely a way to abstract it so we don't have to deal
with it everywhere. Even so, it's still not the worst thing considering
what all bazel has given us so far.