Try using a longer prefix to match
The way this looks to be working is that to work around Windows' lack of
support for symlinks, bazel handles the files differently. It puts a
single `MANIFEST` file in the runfiles directory. That file is supposed
to be used to look up locations of other files that should be in the
runfiles.

`rules_haskell` provides a package we depend on to look up the runfiles
in the tests. It works by attempting to find the file if it exists, or
the directory if it exists. All of this is fine. The issue comes when we
choose what it is that we look up.

When we try to lookup just the workspace ("com_gitlab_joneshf_purty"),
we end up with the first result that matches as the directory where the
runfiles should be. Unfortunately, some of the runfiles point to
different locations. What we end up getting doesn't necessarily work for
the runfiles we need.

A snippet of the `MANIFEST` file on AppVeyor is:

```
...
com_gitlab_joneshf_purty/external/stackage/uuid-types-1.0.3/_install/uuid-types-1.0.3_data C:/users/appveyor/_bazel_appveyor/5rhby5jo/execroot/com_gitlab_joneshf_purty/bazel-out/x64_windows-fastbuild/bin/external/stackage/uuid-types-1.0.3/_install/uuid-types-1.0.3_data
com_gitlab_joneshf_purty/external/stackage/vector-0.12.0.3/_install/vector-0.12.0.3_data C:/users/appveyor/_bazel_appveyor/5rhby5jo/execroot/com_gitlab_joneshf_purty/bazel-out/x64_windows-fastbuild/bin/external/stackage/vector-0.12.0.3/_install/vector-0.12.0.3_data
com_gitlab_joneshf_purty/test/golden/files/formatted/Ado.purs C:/projects/purty/test/golden/files/formatted/Ado.purs
com_gitlab_joneshf_purty/test/golden/files/formatted/Array.purs C:/projects/purty/test/golden/files/formatted/Array.purs
...
```

Two of the entries point to `C:/users/appveyor/_bazel_appveyor/...` and
two of the entries point to `C:/projects/purty/...`.

Depending on the order of the `MANIFEST` file, we would get a different
result for looking up the `rlocation` of `com_gitlab_joneshf_purty`. It
might be
`C:/users/appveyor/_bazel_appveyor/5rhby5jo/execroot/com_gitlab_joneshf_purty/bazel-out/x64_windows-fastbuild/bin/`
or it might be `C:/projects/purty/`. The former doesn't have the golden
files we want, but the latter does.

On Linux and macOS, it doesn't have the same issue for two reasons:
1. The `MANIFEST` file isn't used to find files with `rlocation`.
    On Linux and macOS, `rlocation` uses the file paths directly (which
    happen to be symlinks). It doesn't look in the `MANIFEST` map.
2. Even if the `MANIFEST` file were used, the files would still exist.
    Since Linux and macOS support symlinks, bazel symlinks the files
    into the runfiles directory. If `rlocation` used the same logic for
    Linux and macOS that it does on Windows (looking up files in the
    `MANIFEST`), it would also point to the "wrong" location for some
    files. But, it would work out anyway because that "wrong" location
    would have the files (symlinks) anyway.

The way we address this situation is by using a longer prefix to find.
This should mean that we don't match a location that shouldn't exist,
but rather should match the directory for the golden files.

Given how much work this was, we might want to figure a way to take the
location of files from the outside and let it deal with figuring out
where the files exist.