Move all the CIs over to coverage
It looks like if we run `bazel coverage //...` it will run all the
coverage for all the test targets. If there are targets that aren't for
tests, it looks like it will still build them. So, it seems like we can
build all the artifacts in one go.

Hopefully, there's not some thing that doesn't actually work about this.

Relatedly, now that we are running tests through bazel, we can drop most
of the stack infrastructure. We still have it for now because we're
building `ormolu` and `weeder`. We should figure out a way to deal with them.