Tag the `sh_test` with more metadata
Looks like the `sh_test` is failing because coverage requires the JVM to
be setup. Since we don't want to deal with that on every environment
that isn't already setup, we tag it and filter out things where we need
to.

We also change the bash tag to be more specific and update usages.