Don't run coverage on TravisCI
TravisCI is what we use to build the actual artifacts. If we naively run
coverage, we'll build binaries that output `purty.tix` whenever they're
used.

There's probably a way to build the binaries first for their use as
artifacts and go back and do similar for coverage. Until we get to one
CI, we don't really need to worry about this.