Also cache bazel repository cache on macOS
Bazel uses different locations for the repository cache on different
platforms.

It would be nice to use the `--repository_cache` flag and have a single
place for all platforms, but it doesn't seem to work locally (which is a
linux box). So, it's unclear if it'd work on other platforms.