Cache the stack pantry
The way `rules_haskell` works with the `stack_snapshot` is that it uses
`stack` directly underneath to download dependencies. There's some
intricacies in how things are put together which means that the `stack`
stuff isn't cached by `bazel`:
https://github.com/tweag/rules_haskell/issues/1112. To get around
this (hopefully), we cache the pantry.

According to a comment in a different issue:
https://github.com/tweag/rules_haskell/issues/1167#issuecomment-561630543,
the `STACK_ROOT` environment variable is honored by the `rules_haskell`
setup. We use that so we don't have to muck around in the home directory
too much.

It'd be nice to be done with `stack` once and for all. Maybe we should
try and use something other than `stack_snapshot`?