Use toolchain instead of stack
GHC ships with a bunch of packages already. We don't have to specify the
package as coming from stack because it's always available. Instead, we
can use the `haskell_toolchain_library` rule to make these packages
available.

We don't have a whole lot of dependencies, so it doesn't make much
difference. But, it seems like good practice.