Move `happy` build file to the `tools` directory
We're about to need more build files. So we don't clutter the top-level,
we dump them in the `tools` directory.

This feels a bit off because we don't use `happy` as a tool; it's a
transitive dependency. But, it's better to have it out of the top-level
than to stymie on where to put it.