Setup `ormolu` in `bazel`
We have to do a little weird thing here since `bazel` doesn't work like
`make`. We want to be able to format files and since `bazel` is building
`ormolu` for us, we want to use its knowledge to know which Haskell
files exist and how to format them. Unlike `make`, `bazel` can't have
the input and output of a rule be the same file. So we have to
circumvent that.

We write a `sh_binary` target for this. The `bash` script then hops back
to the actual directory in order to format files in place with `ormolu`.
It's a bit of a roundabout way to get things going, but it seems to
work.

Since we don't care about this script running on all environments, we
just write a `bash` script and call it a day. We'll only be formatting
locally (on Linux), so it shouldn't be much of an issue. Assuming
nothing tries to run this by accident.