Inline sources
It's recommended that we don't attempt to abstract for the sake of
removing syntactic duplication. This is one of those reasons why. We're
about to do different things with `lib/Version.hs`. Each approach we
took turned out worse than the last.

Instead of dealing with that all for the sake of not mentioning these
sources more than once, we inline them where they belong and move on
with our lives.