Manage `purty` version in `bazel`
We want `bazel` to use the version in other targets.

For instance, all of the bintray stuff requires the version. We've been
using `make` to manage it, but we want to push more into `bazel`. We
can't easily do that when the version number is in a file. `bazel`
doesn't allow arbitrary I/O wherever (only specific phases).

We don't necessarily need the version number to be in a file in order
for our build to work appropriately. It was just the method we used at
the time. Since `bazel` can template files, we use it to template out
the `Version` module. This is no worse than using `file-embed` to read
the file at compile time and dump its contents into the
`Version.version` value.

The impetus for templating rather than continuing to use `file-embed` is
primarily because there's not an easy way to use a generated file with
`rules_haskell`: https://github.com/tweag/rules_haskell/issues/1296.
It's a bit unfortunate we cannot use `file-embed` anymore, but it's
really kind of a questionable package to use anyway. Six of one, as they
say.

In order to keep the rest of the build sane, we do actually generate the
`version/purty` file from the moved version like we had before. The
tests still pass, which pretty much means we didn't break the version
behavior!

Finally, we update the documentation to be accurate for this new world.
We haven't been great about keeping this stuff up to date. We should try
to find a better way to do that.