Use `bazel` to create tarball
Now that the `bazel` setup can build the tarball, we use it.

This lets us get rid of a couple of rules. We also drop the dependency
on 7zip in the AppVeyor build. It was provided in the environment, but
it's always better to not have a dependency if there's no adverse
effects.

We next need to figure out the `bintray.json` stuff.