Don't handle platforms differently
We already assume POSIX tooling for all platforms.

We were trying real hard to make sure the Windows stuff stayed sane, but
it doesn't really matter because we were expecting both `curl` and
`touch` to be available. Neither of these are standard Windows programs.

Instead of keeping up this distinction, we just assume standard POSIX
tooling is available, and go from there. Since AppVeyor seems to provide
images that satisfy that, we look to be okay.