Remove `templated` prefix
Looks like we don't need to specify this if we're using rules.

When we were first writing this stuff, we were doing it as a
`write_file` rule with the contents in the `BUILD.bazel` file. It
appeared we had to write out files in a different directory so as not to
conflict. We switch to a rule and using `actions.declare_file`.

Things seem to be a little different if we're using rules. We don't seem
to have an issue with using the same path. So, we can drop the
`templated` prefix we had on our files.