Make a package in `test/golden`
We want to take over what's happening in `test/golden`.

We've been doing a bunch of Haskell stuff in `test/golden`, but none of
this really needs to be written in Haskell. The golden tests should be
running with the binary. The file used test don't need to be written in
Haskell.

In order to start taking over that stuff, we make a package so we can
start building out stuff without muddying up the top-level package.

We ended up having to adjust a few of our macros to work in this new
world. It's nice that these sorts of things are caught by the build
system, instead of failing silently.