Support passing in `bazel` config to `make`
We want CI to be able to use `make` instead of `bazel` directly.

One of the issues with having multiple CI services is that we sometimes
forget to update each of the configurations. To make things a little
easier there, we want to limit how much we have to modify the
configurations. To make that a reality, we have to be able to pass
arguments to `bazel`. We add an argument we can pass to `bazel`, and let
callers modify it.

It would be ideal if the `bazel` build knew what to run and not to run
instead of having to configure it. We should probably look into that.