Fix `make format`
The `//test/golden:format-ormolu` target no longer exists.

It would be nice to have this in `bazel` so we wouldn't have to deal
with this stuff not lining up. The reason it's not is that we can't run
multiple targets in a single command.

If we can figure out some other way to do this, we should try to move
over to that approach.