Write `actions.expand_template` once
We don't want to have multiple `actions.expand_template` calls.

We have to write `actions.expand_template` ourselves because there's no
simple way to use it. But we don't have to write it multiple times.
Instead, we define one place that does the templating, and use it in
other areas.

We cannot get rid of the travis-specific stuff yet, because it doesn't
allow dicts with labels as values (so the expansion of files doesn't
work). But that travis stuff is a small price to pay.