Move `purty` library to its own package
We can move the `purty` library to its own package now.

Moving everything to packages makes `bazel` work a little better. This
is mostly because crossing directories is not straight-forward with
`bazel`. We end up doing some hacks to work with it.

In this case, when we had `//:purty`, we had to hack around stuff with
`src_strip_prefix`. Now that have `//lib:purty`, we can drop the hack.
There's other little things like that hack that we end up doing
throughout the build system that go away once everything is in its own
package.