Move `purty` binary to its own package
We move the creation of `purty` to its own package.

This is similar motivation to what we did for the `purty` library. The
change in target shows another hack we had to make when we had a single
package. With a single package, we cannot have multiple targets with the
same name. So, we couldn't name the `purty` binary and library both
`"purty"`. Turns out, there's something in `rules_haskell` for this
case. The alternative we took was to rename the binary. Now that we
don't have conflicts, we can name the binary `"purty"`.

It's also pretty nice that the change in targets worked out where the
golden tests had to be updated for the change. We probably would have
forgotten this in `make`.