Move AppVeyor-specific tarball to a package
We want the AppVeyor-specific stuff to be in its own package.

This is a really nice benefit of using `bazel`. It becomes trivial to
write stuff specific for a single use. The triviality is that the
targets are checked in the build system. Things outside of the build are
a bit harder to deal with (like us needing to remember to update the
`appveyor.yml` file).